refactor: extract WorkspaceState settings into an extension file
Move the settings lifecycle (profile/relays/notifications/privacy load and
save, key-package management, audit-log settings and upload, relay-draft
editing and validation) into WorkspaceState+Settings.swift. Pure code
movement, no behavior change. Build green; all 208 tests pass.
